+++ beer-blocks.php
@@ -5,8 +5,8 @@
  * Description: Beer Blocks extend the block editor with advanced and powerful blocks that help you build beautiful websites faster.
  * Requires at least: 5.6
  * Requires PHP: 7.3
- * Version: 1.1.1
- * Author: CoDevelopers
+ * Version: 1.2.0
+ * Author: codevelopers
  * Author URI: https://codevelopers.tech
  * License: GPL-2.0-or-later
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html