+++ src/store/selectors.js
@@ -1,9 +1,0 @@
-export const selectedFonts = (state) => state.googleFonts.selectedFonts;
-export const selectedFontsLoading = (state) =>
-	state.googleFonts.selectedFontsLoading;
-export const selectedFontsError = (state) =>
-	state.googleFonts.selectedFontsError;
-export const updateSelectedFontsLoading = (state) =>
-	state.googleFonts.updateSelectedFontsLoading;
-export const updateSelectedFontsError = (state) =>
-	state.googleFonts.updateSelectedFontsError;